phase one: move specific task implementations into submodule
