+++ nexus/src/app/background/tasks/bfd.rs
@@ -6,7 +6,7 @@
 //! (BFD) sessions.
 
 use crate::app::{
-    background::networking::build_mgd_clients, map_switch_zone_addrs,
+    background::tasks::networking::build_mgd_clients, map_switch_zone_addrs,
 };
 
 use super::common::BackgroundTask;
