+++ nexus/src/app/background/tasks/metrics_producer_gc.rs
